# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Workout API",
    "description": "A workout tracking API built with FastAPI",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/auth/google": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Initiate Google OAuth flow",
        "description": "Start Google OAuth authentication process and return authorization URL",
        "operationId": "initiate_google_oauth_api_v1_auth_google_post",
        "parameters": [
          {
            "name": "redirect_url",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Post-auth redirect URL",
              "title": "Redirect Url"
            },
            "description": "Post-auth redirect URL"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OAuthInitiateResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/google/callback": {
      "get": {
        "tags": [
          "Authentication"
        ],
        "summary": "Handle Google OAuth callback",
        "description": "Process Google OAuth callback and return user with JWT tokens",
        "operationId": "google_oauth_callback_api_v1_auth_google_callback_get",
        "parameters": [
          {
            "name": "code",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Authorization code from Google",
              "title": "Code"
            },
            "description": "Authorization code from Google"
          },
          {
            "name": "state",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "OAuth state parameter",
              "title": "State"
            },
            "description": "OAuth state parameter"
          },
          {
            "name": "error",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "OAuth error",
              "title": "Error"
            },
            "description": "OAuth error"
          },
          {
            "name": "error_description",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "OAuth error description",
              "title": "Error Description"
            },
            "description": "OAuth error description"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthenticationResponse"
                }
              }
            }
          },
          "400": {
            "description": "OAuth error or invalid state",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Authentication failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/refresh": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Refresh access token",
        "description": "Use refresh token to get new access token",
        "operationId": "refresh_token_api_v1_auth_refresh_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TokenRefreshRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenRefreshResponse"
                }
              }
            }
          },
          "401": {
            "description": "Invalid refresh token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Logout user",
        "description": "Logout current user (client-side token invalidation)",
        "operationId": "logout_api_v1_auth_logout_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LogoutResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/auth/me": {
      "get": {
        "tags": [
          "Authentication"
        ],
        "summary": "Get current session info",
        "description": "Get current user session information",
        "operationId": "get_session_info_api_v1_auth_me_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionInfoResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/auth/validate": {
      "get": {
        "tags": [
          "Authentication"
        ],
        "summary": "Validate token",
        "description": "Validate JWT token and return token information",
        "operationId": "validate_token_api_v1_auth_validate_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenValidationResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "AuthErrorResponse": {
        "properties": {
          "error": {
            "type": "string",
            "title": "Error",
            "description": "Error type"
          },
          "error_description": {
            "type": "string",
            "title": "Error Description",
            "description": "Detailed error description"
          },
          "error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code",
            "description": "Internal error code"
          }
        },
        "type": "object",
        "required": [
          "error",
          "error_description"
        ],
        "title": "AuthErrorResponse",
        "description": "Authentication error response."
      },
      "AuthenticationResponse": {
        "properties": {
          "user": {
            "$ref": "#/components/schemas/UserProfileResponse",
            "description": "User profile information"
          },
          "tokens": {
            "$ref": "#/components/schemas/TokenResponse",
            "description": "JWT token pair"
          }
        },
        "type": "object",
        "required": [
          "user",
          "tokens"
        ],
        "title": "AuthenticationResponse",
        "description": "Complete authentication response with user and tokens."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "LogoutResponse": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Logout message",
            "default": "Successfully logged out"
          },
          "logged_out": {
            "type": "boolean",
            "title": "Logged Out",
            "description": "Logout status",
            "default": true
          }
        },
        "type": "object",
        "title": "LogoutResponse",
        "description": "Logout response."
      },
      "OAuthInitiateResponse": {
        "properties": {
          "authorization_url": {
            "type": "string",
            "maxLength": 2083,
            "minLength": 1,
            "format": "uri",
            "title": "Authorization Url",
            "description": "Google OAuth authorization URL"
          },
          "state": {
            "type": "string",
            "title": "State",
            "description": "OAuth state parameter for CSRF protection"
          }
        },
        "type": "object",
        "required": [
          "authorization_url",
          "state"
        ],
        "title": "OAuthInitiateResponse",
        "description": "Response with OAuth authorization URL."
      },
...
      "SessionInfoResponse": {
        "properties": {
          "authenticated": {
            "type": "boolean",
            "title": "Authenticated",
            "description": "Whether user is authenticated"
          },
          "user": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/UserProfileResponse"
              },
              {
                "type": "null"
              }
            ],
            "description": "User profile if authenticated"
          },
          "session_expires_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Session Expires At",
            "description": "When the session expires"
          },
          "permissions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Permissions",
            "description": "User permissions"
          }
        },
        "type": "object",
        "required": [
          "authenticated"
        ],
        "title": "SessionInfoResponse",
        "description": "Current session information."
      },
...
      "TokenRefreshRequest": {
        "properties": {
          "refresh_token": {
            "type": "string",
            "title": "Refresh Token",
            "description": "Valid refresh token"
          }
        },
        "type": "object",
        "required": [
          "refresh_token"
        ],
        "title": "TokenRefreshRequest",
        "description": "Token refresh request."
      },
      "TokenRefreshResponse": {
        "properties": {
          "access_token": {
            "type": "string",
            "title": "Access Token",
            "description": "New JWT access token"
          },
          "token_type": {
            "type": "string",
            "title": "Token Type",
            "description": "Token type",
            "default": "Bearer"
          },
          "expires_in": {
            "type": "integer",
            "title": "Expires In",
            "description": "Access token expiration in seconds"
          }
        },
        "type": "object",
        "required": [
          "access_token",
          "expires_in"
        ],
        "title": "TokenRefreshResponse",
        "description": "Token refresh response."
      },
      "TokenResponse": {
        "properties": {
          "access_token": {
            "type": "string",
            "title": "Access Token",
            "description": "JWT access token"
          },
          "refresh_token": {
            "type": "string",
            "title": "Refresh Token",
            "description": "JWT refresh token"
          },
          "token_type": {
            "type": "string",
            "title": "Token Type",
            "description": "Token type",
            "default": "Bearer"
          },
          "expires_in": {
            "type": "integer",
            "title": "Expires In",
            "description": "Access token expiration in seconds"
          }
        },
        "type": "object",
        "required": [
          "access_token",
          "refresh_token",
          "expires_in"
        ],
        "title": "TokenResponse",
        "description": "JWT token response."
      },
      "TokenValidationResponse": {
        "properties": {
          "valid": {
            "type": "boolean",
            "title": "Valid",
            "description": "Whether the token is valid"
          },
          "user_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Id",
            "description": "User ID if token is valid"
          },
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email",
            "description": "User email if token is valid"
          },
          "expires_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires At",
            "description": "Token expiration timestamp"
          },
          "token_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Token Type",
            "description": "Type of token"
          }
        },
        "type": "object",
        "required": [
          "valid"
        ],
        "title": "TokenValidationResponse",
        "description": "Token validation response."
      },
      "UserProfileResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id",
            "description": "User ID"
          },
          "email_address": {
            "type": "string",
            "title": "Email Address",
            "description": "User email address"
          },
          "google_id": {
            "type": "string",
            "title": "Google Id",
            "description": "Google OAuth user ID"
          },
          "name": {
            "type": "string",
            "title": "Name",
            "description": "User display name"
          },
          "profile_image_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Profile Image Url",
            "description": "URL to user's profile image"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active",
            "description": "Whether the user account is active"
          },
          "is_admin": {
            "type": "boolean",
            "title": "Is Admin",
            "description": "Whether the user has admin privileges"
          }
        },
        "type": "object",
        "required": [
          "id",
          "email_address",
          "google_id",
          "name",
          "profile_image_url",
          "is_active",
          "is_admin"
        ],
        "title": "UserProfileResponse",
        "description": "User profile response."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}